VK-36 Комментарии - раскрытие комментариев - удаление по правкам - по 15 на страницу

+++ plugins/books/profile/classes/ProfilerService.php
@@ -2,6 +2,7 @@
 
 namespace Books\Profile\Classes;
 
+use Books\Comments\Models\Comment;
 use Books\Profile\Models\Profiler;
 use October\Rain\Database\Model;
 
@@ -9,7 +10,6 @@ class ProfilerService
 {
     public function __construct(protected Model $model)
     {
-
     }
 
     public function add(): void
@@ -19,12 +19,13 @@ class ProfilerService
         $profiler->master()->associate($master);
         $profiler->slave()->associate($this->model);
         $profiler->save();
-
     }
 
     public function remove(): void
     {
+        if ($this->model instanceof Comment) {
+            return;
+        }
         $this->model->profiler()?->delete();
     }
-
 }
